+++ CHANGELOG.md
@@ -4,6 +4,34 @@ All notable changes to the Platfuse Game Engine project will be documented in th
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.x.x] - UNRELEASED
+
+### Added
+
+-   Image layers.
+-   Layer `offset`, `parallaxFactor`, `repeat` and `tint` properties.
+-   WebGL post process.
+
+### Changed
+
+-   N/A
+
+### Deprecated
+
+-   N/A
+
+### Removed
+
+-   N/A
+
+### Fixed
+
+-   N/A
+
+### Security
+
+-   N/A
+
 ## [1.5.0] - 2024-07-19
 
 ### Added